+++ Source/cmake.cxx
@@ -811,8 +811,10 @@ bool cmake::CacheVersionMatches()
   const char* minv = m_CacheManager->GetCacheValue("CMAKE_CACHE_MINOR_VERSION");
   const char* relv = m_CacheManager->GetCacheValue("CMAKE_CACHE_RELEASE_VERSION");
   bool cacheSameCMake = false;
-  if(majv && atoi(majv) == cmMakefile::GetMajorVersion()
-     && minv && atoi(minv) == cmMakefile::GetMinorVersion()
+  if(majv && 
+     atoi(majv) == static_cast<int>(cmMakefile::GetMajorVersion())
+     && minv && 
+     atoi(minv) == static_cast<int>(cmMakefile::GetMinorVersion())
      && relv && (strcmp(relv, cmMakefile::GetReleaseVersion()) == 0))
     {
     cacheSameCMake = true;
